fix: enforce worktree-aware repair and shell policy
