+++ lib/commands/status.js
@@ -501,8 +501,8 @@ function formatStatus(result) {
  * @param {string} projectRoot - Absolute path to project root
  * @returns {object} Context for detectStage
  */
-function buildContext(projectRoot) {
-	const root = projectRoot || process.cwd();
+function _buildContext(_projectRoot) {
+	const root = _projectRoot || process.cwd();
 	let branch = 'master';
 	try {
 		branch = execFileSync('git', ['rev-parse', '--abbrev-ref', 'HEAD'], {
@@ -535,7 +535,7 @@ function buildContext(projectRoot) {
 module.exports = {
 	name: 'status',
 	description: 'Intelligent stage detection with confidence scoring',
-	handler: async (args, flags, projectRoot) => {
+	handler: async (args, flags, _projectRoot) => {
 		const inputs = parseStatusInputs(args, flags);
 		const workflowState = inputs.workflowState
 			? readWorkflowState(inputs.workflowState)
