+++ lib/husky-migration.js
@@ -11,6 +11,65 @@ const fs = require('node:fs');
 const path = require('node:path');
 const { execFileSync, spawnSync } = require('node:child_process');
 
+function readGitDirPointer(projectRoot) {
+  const gitPath = path.join(projectRoot, '.git');
+  if (!fs.existsSync(gitPath)) {
+    return null;
+  }
+
+  try {
+    const gitStat = fs.statSync(gitPath);
+    if (gitStat.isDirectory()) {
+      return gitPath;
+    }
+    if (!gitStat.isFile()) {
+      return null;
+    }
+
+    const pointer = fs.readFileSync(gitPath, 'utf8').trim();
+    if (!pointer.startsWith('gitdir:')) {
+      return null;
+    }
+
+    return path.resolve(projectRoot, pointer.replace(/^gitdir:\s*/, ''));
+  } catch (_err) {
+    return null;
+  }
+}
+
+function hasHooksPathConfigured(projectRoot) {
+  const gitDir = readGitDirPointer(projectRoot);
+  if (!gitDir) {
+    return false;
+  }
+
+  const configCandidates = [path.join(gitDir, 'config')];
+  const commonDirPath = path.join(gitDir, 'commondir');
+  if (fs.existsSync(commonDirPath)) {
+    try {
+      const commonDir = fs.readFileSync(commonDirPath, 'utf8').trim();
+      if (commonDir) {
+        configCandidates.push(path.resolve(gitDir, commonDir, 'config'));
+      }
+    } catch (_err) {
+      // Ignore unreadable commondir files and fall back to direct config candidates.
+    }
+  }
+
+  return configCandidates.some((configPath) => {
+    if (!fs.existsSync(configPath)) {
+      return false;
+    }
+
+    try {
+      const gitConfig = fs.readFileSync(configPath, 'utf8');
+      return /hooksPath\s*=/.test(gitConfig);
+    } catch (_err) {
+      return false;
+    }
+  });
+}
+
 /**
  * Resolve a command to its full path to avoid relying on inherited PATH.
  * Falls back to the command name if resolution fails.
@@ -67,13 +126,39 @@ const KNOWN_PATTERNS = [
  */
 const HUSKY_INTERNAL = new Set(['_', '.gitignore', 'husky.sh']);
 
+function resolveGitDir(projectRoot) {
+  const gitPath = path.join(projectRoot, '.git');
+  if (!fs.existsSync(gitPath)) {
+    return null;
+  }
+
+  try {
+    const stat = fs.statSync(gitPath);
+    if (stat.isDirectory()) {
+      return gitPath;
+    }
+
+    if (stat.isFile()) {
+      const pointer = fs.readFileSync(gitPath, 'utf8').trim();
+      const match = /^gitdir:\s*(.+)$/i.exec(pointer);
+      if (match && match[1]) {
+        return path.resolve(projectRoot, match[1].trim());
+      }
+    }
+  } catch (_err) {
+    return null;
+  }
+
+  return null;
+}
+
 /**
  * Detect whether Husky is installed in a project.
  *
  * @param {string} projectRoot - Absolute path to the project root.
  * @returns {{ found: boolean, huskyDir: string|null, hasHooksPath: boolean }}
  */
-function detectHusky(projectRoot) {
+function _detectHuskyLegacy(projectRoot) {
   const huskyDir = path.join(projectRoot, '.husky');
   const found = fs.existsSync(huskyDir) && fs.statSync(huskyDir).isDirectory();
 
@@ -83,8 +168,9 @@ function detectHusky(projectRoot) {
 
   // Check for core.hooksPath in .git/config
   let hasHooksPath = false;
-  const gitConfigPath = path.join(projectRoot, '.git', 'config');
-  if (fs.existsSync(gitConfigPath)) {
+  const gitDir = resolveGitDir(projectRoot);
+  const gitConfigPath = gitDir ? path.join(gitDir, 'config') : null;
+  if (gitConfigPath && fs.existsSync(gitConfigPath)) {
     try {
       const gitConfig = fs.readFileSync(gitConfigPath, 'utf8');
       hasHooksPath = /hooksPath\s*=/.test(gitConfig);
@@ -94,6 +180,17 @@ function detectHusky(projectRoot) {
   }
 
   return { found: true, huskyDir, hasHooksPath };
+}
+
+function detectHusky(projectRoot) {
+  const huskyDir = path.join(projectRoot, '.husky');
+  const found = fs.existsSync(huskyDir) && fs.statSync(huskyDir).isDirectory();
+
+  if (!found) {
+    return { found: false, huskyDir: null, hasHooksPath: false };
+  }
+
+  return { found: true, huskyDir, hasHooksPath: hasHooksPathConfigured(projectRoot) };
 }
 
 /**
