+++ lib/workflow/enforce-stage.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const { repairWorkflowRuntimeAssets } = require('../commands/setup');
 const { checkRuntimeHealth } = require('../runtime-health');
 const { getAllowedTransitions, normalizeStageId } = require('./stages');
 const { normalizeOverrideRecord, readWorkflowState } = require('./state');
@@ -38,13 +39,29 @@ function formatDiagnostics(diagnostics = []) {
     .join('; ');
 }
 
-async function enforceStageEntry({ commandName, flags = {}, projectRoot, workflowState, health } = {}) {
+async function enforceStageEntry({ commandName, flags = {}, projectRoot, workflowState, health, repairRuntime } = {}) {
   const stageId = normalizeStageId(commandName);
   if (!stageId) {
     return { allowed: true };
   }
 
-  const runtimeHealth = health || checkRuntimeHealth(projectRoot);
+  if (projectRoot) {
+    repairWorkflowRuntimeAssets(projectRoot);
+  }
+
+  let runtimeHealth = health || checkRuntimeHealth(projectRoot);
+  if (runtimeHealth.hardStop && typeof repairRuntime === 'function') {
+    const repairedHealth = await repairRuntime({
+      commandName,
+      flags,
+      projectRoot,
+      workflowState,
+      health: runtimeHealth,
+    });
+    if (repairedHealth) {
+      runtimeHealth = repairedHealth;
+    }
+  }
   if (runtimeHealth.hardStop) {
     throw new Error(`Stage ${stageId} blocked by runtime prerequisites: ${formatDiagnostics(runtimeHealth.diagnostics)}`);
   }
